+++ react/RatingSummary.tsx
@@ -22,13 +22,24 @@ const RatingSummary: FunctionComponent<
   const { product } = useContext(ProductContext)
 
   const writeReviewLink =
-    props.data && props.data.getConfig
+    props.data && props.data.getConfig && product
       ? `/new-review?pr_page_id=${
           product[props.data.getConfig.uniqueId]
         }&pr_merchant_id=${props.data.getConfig.merchantId}&pr_api_key=${
           props.data.getConfig.appKey
         }&pr_merchant_group_id=${props.data.getConfig.merchantGroupId}`
       : ''
+
+  if (!product) {
+    return (
+      <Summary
+        loading
+        writeReviewLink={writeReviewLink}
+        rating={0}
+        numberOfReviews={0}
+      />
+    )
+  }
 
   return (
     <Query
@@ -146,14 +157,17 @@ const Summary: FunctionComponent<SummaryProps> = ({
           })}
         </div>
       </div>
-      <span className="review__rating--average mr4 dib c-muted-2 t-body f6-s">
+      <a
+        href="#all-reviews"
+        className="review__rating--average mr4 dib c-muted-2 t-body f6-s"
+      >
         (
         <FormattedMessage
           id="store/power-reviews.numberOfReviews"
           values={{ number: numberOfReviews }}
         />
         )
-      </span>
+      </a>
       {!loading ? (
         <Link className="dib c-on-base t-body f6-s" to={writeReviewLink}>
           <FormattedMessage id="store/power-reviews.writeAReview" />